Add Image model and a file upload controller to the API.
This commit introduces the Image model which is related to the Post model. It allows the attachment of images to blog posts imported from Headway. The images are downloaded from the source and stored using ActiveStorage.

In addition, a new file upload controller () has been created. This will handle the attachment of downloaded images to the respective blog posts.

Note: This enhancement enables the migration script to handle posts with images seamlessly, ensuring a complete and accurate import of data from Headway to Changepack.

Tests and the respective factory_bot factories for the Image model are also added to ensure correct functionality.